+++ cdap-examples/LogAnalysis/src/main/java/co/cask/cdap/examples/loganalysis/ApacheAccessLog.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package co.cask.cdap.examples.sparkloganalysis;
+package co.cask.cdap.examples.loganalysis;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
